refactoring

See merge request compsci308_2020spring/parser_parser_team17!142

+++ src/backEnd/commands/ClearScreen.java
@@ -24,7 +24,7 @@ public class ClearScreen extends Command {
     return ret;
   }
   @Override
-  public int repeatCom() {
+  public double repeatCom() {
     return super.repeatCom();
   }
 }
