+++ src/backEnd/commands/Left.java
@@ -24,7 +24,7 @@ public class Left extends Command {
   }
 
   @Override
-  public int repeatCom() {
+  public double repeatCom() {
     return super.repeatCom();
   }
 }
