+++ src/backEnd/commands/PenDown.java
@@ -23,7 +23,7 @@ public class PenDown extends Command {
   }
 
   @Override
-  public int repeatCom() {
+  public double repeatCom() {
     return super.repeatCom();
   }
 }
