+++ src/backEnd/commands/XCoordinate.java
@@ -15,7 +15,6 @@ public class XCoordinate extends Command {
     public XCoordinate(LinkedList<String> varargs, Control control) {
         super(varargs, control);
         super.numberOfArgs = NUMARGS;
-
         returnXCord = control.getTurtleRelativeXPos();
     }
 
@@ -25,7 +24,7 @@ public class XCoordinate extends Command {
     }
 
     @Override
-    public int repeatCom() {
+    public double repeatCom() {
         return super.repeatCom();
     }
 }
